chore: Move where we export the hook
